requirements.txt
# FastAPI and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0

# CORS
python-multipart>=0.0.6

# Pydantic email validation
email-validator>=2.1.0

# Celery and task queue
celery>=5.3.0
redis>=5.0.0

# AWS and encryption
cryptography>=41.0.0
boto3>=1.28.0
python-dotenv>=1.0.0
